Split vnf_p7.h into vnf_p7_lte and vnf_p7_nr header files
