+++ nfapi/open-nFAPI/vnf/src/vnf_p7_nr.c
@@ -8,7 +8,7 @@
 #include <stdio.h>
 #include <string.h>
 
-#include "vnf_p7.h"
+#include "vnf_p7_nr.h"
 #include "nr_fapi_p7_utils.h"
 
 #define SYNC_CYCLE_COUNT 2
